Refactor profiles/views.py into a module folder, in preparation for an additional OrganizationProfile Views
