+++ pulseapi/profiles/views/__init__.py
@@ -1,0 +1,1 @@
+from profiles.views.userprofile import *
